refactor: colocalize environment context with config domain
